RW data displays in UI
Partial transition from 'friends' semantics to 'RW' semantics.
DataGrid on the Friends tab now updates with data from the inspector
