+++ client/GameData/RWs/DataTargetProfiles.cs
@@ -7,18 +7,16 @@ using System.Text;
 using System.Threading.Tasks;
 using System.ComponentModel;
 
-namespace FFRKInspector.GameData.Friends
+namespace FFRKInspector.GameData.RWs
 {
-    class DataRelatedRW
+    class DataTargetProfile
     {
-        //user relation and target profile data combined here. figure out how to default null missing data until target profile is added?
-        //  and array of this to be stored in GameState with all followers AND followees, UI to differentiate by relation_status value (1 is follower, 2 is followee, 3 is mutual)
-
-        ///relation_status = 0 is no relation (from battle selection), 1 is follower, 2 is followee, 3 is mutual
-        [JsonProperty("relation_status")]
+        //relation_status = 0 is no relation (from battle selection), 1 is follower, 2 is followee, 3 is mutual
+        [JsonProperty("relation_status", DefaultValueHandling = DefaultValueHandling.Populate)]
+        [DefaultValue(0)]
         public byte Status;
 
-        [JsonProperty("target_user_id")]
+        [JsonProperty("user_id")]
         public ulong UserID;
 
         ///nickname = player name (for identification in app list)
@@ -65,31 +63,5 @@ namespace FFRKInspector.GameData.Friends
         [JsonProperty(PropertyName = "supporter_buddy_spd", DefaultValueHandling = DefaultValueHandling.Populate)]
         [DefaultValue(0)]
         public UInt16 Speed;
-
-        public void UpdateRWData(DataRelatedRW NewRW)
-        {
-            this.Accuracy = NewRW.Accuracy;
-            this.Attack = NewRW.Attack;
-            this.Magic = NewRW.Magic;
-            this.Mind = NewRW.Mind;
-            this.NickName = NewRW.NickName;
-            this.SBName = NewRW.SBName;
-            this.SBType = NewRW.SBType;
-            this.Slogan = NewRW.Slogan;
-            this.Speed = NewRW.Speed;
-        }
-
-        public void UpdateRWData(DataTargetProfile NewProfile)
-        {
-            this.Accuracy = NewProfile.Accuracy;
-            this.Attack = NewProfile.Attack;
-            this.Magic = NewProfile.Magic;
-            this.Mind = NewProfile.Mind;
-            this.NickName = NewProfile.NickName;
-            this.SBName = NewProfile.SBName;
-            this.SBType = NewProfile.SBType;
-            this.Slogan = NewProfile.Slogan;
-            this.Speed = NewProfile.Speed;
-        }
     }
 }
